Divide transforms into sam's and others
