+++ otx/algorithms/visual_prompting/adapters/pytorch_lightning/datasets/pipelines/__init__.py
@@ -3,6 +3,7 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .transforms import MultipleInputsCompose, Pad, ResizeLongestSide, collate_fn
+from .sam_transforms import ResizeLongestSide
+from .transforms import MultipleInputsCompose, Pad, collate_fn
 
 __all__ = ["ResizeLongestSide", "collate_fn", "MultipleInputsCompose", "Pad"]
